+++ eras/conway/impl/test/Test/Cardano/Ledger/Conway/GoldenSpec.hs
@@ -7,6 +7,7 @@ module Test.Cardano.Ledger.Conway.GoldenSpec (spec) where
 import Cardano.Ledger.Conway (ConwayEra)
 import Paths_cardano_ledger_conway (getDataFileName)
 import Test.Cardano.Ledger.Common
+import qualified Test.Cardano.Ledger.Conway.Binary.Golden as Golden
 import Test.Cardano.Ledger.Conway.Era ()
 import Test.Cardano.Ledger.Core.JSON (goldenJsonPParamsSpec, goldenJsonPParamsUpdateSpec)
 
@@ -17,3 +18,4 @@ spec =
       goldenJsonPParamsSpec @ConwayEra
     beforeAll (getDataFileName "golden/pparams-update.json") $
       goldenJsonPParamsUpdateSpec @ConwayEra
+    describe "Conway era golden tests" $ Golden.spec @ConwayEra