+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/Binary/Golden.hs
@@ -9,66 +9,117 @@
 
 module Test.Cardano.Ledger.Dijkstra.Binary.Golden (
   spec,
+  module Test.Cardano.Ledger.Conway.Binary.Golden,
 ) where
 
 import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusTxInfo, SupportedLanguage (..))
 import Cardano.Ledger.Alonzo.Scripts (plutusScriptBinary)
 import Cardano.Ledger.Alonzo.TxWits (Redeemers)
 import Cardano.Ledger.BaseTypes (Version)
-import Cardano.Ledger.Binary (DecoderError (..), DeserialiseFailure (..), Tokens (..))
-import Cardano.Ledger.Coin (Coin (..))
-import Cardano.Ledger.Conway.TxCert (Delegatee (..))
-import Cardano.Ledger.Credential (Credential (..))
+import Cardano.Ledger.Binary (Annotator, DecoderError (..), DeserialiseFailure (..), Tokens (..))
+import qualified Cardano.Ledger.Binary as Binary
 import Cardano.Ledger.Dijkstra.Core (
-  EraTxBody (..),
-  EraTxOut (..),
   EraTxWits (..),
-  TxLevel (..),
-  eraProtVerLow,
-  pattern DelegTxCert,
  )
 import Cardano.Ledger.Plutus (SLanguage (..))
-import Cardano.Ledger.TxIn (TxIn (..))
-import qualified Data.Set as Set
+import Data.Data (Proxy (..))
 import Test.Cardano.Ledger.Alonzo.Arbitrary (alwaysSucceedsLang)
 import Test.Cardano.Ledger.Binary.Plain.Golden (Enc (..))
 import Test.Cardano.Ledger.Common (Spec, describe, it)
-import Test.Cardano.Ledger.Conway.Binary.Golden (expectDecoderFailureAnn, listRedeemersEnc)
-import Test.Cardano.Ledger.Core.KeyPair (mkKeyHash, mkKeyPair, mkWitnessVKey)
+import Test.Cardano.Ledger.Conway.Binary.Golden hiding (spec)
+import Test.Cardano.Ledger.Core.KeyPair (mkKeyPair, mkWitnessVKey)
 import Test.Cardano.Ledger.Core.Utils (mkDummySafeHash)
 import Test.Cardano.Ledger.Dijkstra.Era (DijkstraEraTest)
+import Test.Cardano.Ledger.Imp.Common (forEachEraVersion)
 
 spec :: forall era. DijkstraEraTest era => Spec
-spec = describe "Golden" $ do
-  goldenListRedeemersDisallowed @era
-  goldenDuplicateCertsDisallowed @era
+spec = describe "Golden" . forEachEraVersion @era $ \version -> do
+  describe "Redeemers" $ do
+    goldenListRedeemersDisallowed @era version
+  describe "TxCert" $ do
+    conwayDecodeDuplicateDelegCertFails @era version
   describe "TxWits" $ do
-    goldenDuplicateVKeyWitsDisallowed @era
-    goldenDuplicateNativeScriptsDisallowed @era
-    goldenDuplicatePlutusScriptsDisallowed @era SPlutusV1
-    goldenDuplicatePlutusScriptsDisallowed @era SPlutusV2
-    goldenDuplicatePlutusScriptsDisallowed @era SPlutusV3
-    goldenDuplicatePlutusDataDisallowed @era
+    goldenDuplicateVKeyWitsDisallowed @era version
+    goldenDuplicateNativeScriptsDisallowed @era version
+    goldenDuplicatePlutusScriptsDisallowed @era version SPlutusV1
+    goldenDuplicatePlutusScriptsDisallowed @era version SPlutusV2
+    goldenDuplicatePlutusScriptsDisallowed @era version SPlutusV3
+    goldenDuplicatePlutusDataDisallowed @era version
+    goldenEmptyFields @era version
 
-duplicateCertsTx :: forall era. DijkstraEraTest era => Version -> Enc
-duplicateCertsTx v =
-  mconcat
-    [ E $ TkMapLen 4
-    , Em [E @Int 0, Ev v $ Set.empty @TxIn]
-    , Em [E @Int 1, Ev v $ [] @(TxOut era)]
-    , Em [E @Int 2, E $ Coin 0]
-    , Em
-        [ E @Int 4
-        , Em
-            [ E $ TkTag 258
-            , E $ TkListLen 2
-            , Ev v cert
-            , Ev v cert
-            ]
-        ]
-    ]
-  where
-    cert = DelegTxCert @era (KeyHashObj (mkKeyHash 0)) (DelegStake (mkKeyHash 1))
+goldenEmptyFields :: forall era. DijkstraEraTest era => Version -> Spec
+goldenEmptyFields version =
+  describe "Empty fields not allowed" $ do
+    describe "Untagged" $ do
+      let
+        expectFailureOnEmptyField k =
+          expectDecoderFailureAnn @(TxWits era) version (witsEmptyField k)
+      it "addrTxWits" . expectFailureOnEmptyField 0 $
+        DecoderErrorDeserialiseFailure
+          (Binary.label $ Proxy @(Annotator (TxWits era)))
+          (DeserialiseFailure 4 "Set cannot be empty")
+      it "nativeScripts" . expectFailureOnEmptyField 1 $
+        DecoderErrorCustom "Annotator" "Empty script Set is not allowed"
+      it "bootstrapWitness" . expectFailureOnEmptyField 2 $
+        DecoderErrorDeserialiseFailure
+          (Binary.label $ Proxy @(Annotator (TxWits era)))
+          (DeserialiseFailure 4 "Set cannot be empty")
+      it "plutusV1Script" . expectFailureOnEmptyField 3 $
+        DecoderErrorDeserialiseFailure
+          "Annotator (MemoBytes (AlonzoTxWitsRaw DijkstraEra))"
+          (DeserialiseFailure 4 "Empty list of scripts is not allowed")
+      it "plutusData" . expectFailureOnEmptyField 4 $
+        DecoderErrorCustom "Annotator" "Empty script Set is not allowed"
+      it "redeemers" . expectFailureOnEmptyField 5 $
+        DecoderErrorDeserialiseFailure
+          "Annotator (MemoBytes (AlonzoTxWitsRaw DijkstraEra))"
+          (DeserialiseFailure 2 "List encoding of redeemers not supported starting with PV 12")
+      it "plutusV2Script" . expectFailureOnEmptyField 6 $
+        DecoderErrorDeserialiseFailure
+          "Annotator (MemoBytes (AlonzoTxWitsRaw DijkstraEra))"
+          (DeserialiseFailure 4 "Empty list of scripts is not allowed")
+      it "plutusV3Script" . expectFailureOnEmptyField 7 $
+        DecoderErrorDeserialiseFailure
+          "Annotator (MemoBytes (AlonzoTxWitsRaw DijkstraEra))"
+          (DeserialiseFailure 4 "Empty list of scripts is not allowed")
+      -- TODO replace this with `plutusV4Script` once that is added
+      it "8th field" . expectFailureOnEmptyField 8 $
+        DecoderErrorDeserialiseFailure
+          (Binary.label $ Proxy @(Annotator (TxWits era)))
+          (DeserialiseFailure 2 "An error occured while decoding (Int,Void) not a valid key:.\nError: 8")
+    describe "Tagged" $ do
+      let
+        expectFailureOnEmptyField k =
+          expectDecoderFailureAnn @(TxWits era) version (witsEmptyFieldWithTag k)
+      it "addrTxWits" . expectFailureOnEmptyField 0 $
+        DecoderErrorDeserialiseFailure
+          (Binary.label $ Proxy @(Annotator (TxWits era)))
+          (DeserialiseFailure 7 "Set cannot be empty")
+      it "nativeScripts" . expectFailureOnEmptyField 1 $
+        DecoderErrorCustom "Annotator" "Empty script Set is not allowed"
+      it "bootstrapWitness" . expectFailureOnEmptyField 2 $
+        DecoderErrorDeserialiseFailure
+          (Binary.label $ Proxy @(Annotator (TxWits era)))
+          (DeserialiseFailure 7 "Set cannot be empty")
+      it "plutusV1Script" . expectFailureOnEmptyField 3 $
+        DecoderErrorDeserialiseFailure
+          (Binary.label $ Proxy @(Annotator (TxWits era)))
+          (DeserialiseFailure 7 "Empty list of scripts is not allowed")
+      it "plutusData" . expectFailureOnEmptyField 4 $
+        DecoderErrorCustom "Annotator" "Empty script Set is not allowed"
+      it "plutusV2Script" . expectFailureOnEmptyField 6 $
+        DecoderErrorDeserialiseFailure
+          (Binary.label $ Proxy @(Annotator (TxWits era)))
+          (DeserialiseFailure 7 "Empty list of scripts is not allowed")
+      it "plutusV3Script" . expectFailureOnEmptyField 7 $
+        DecoderErrorDeserialiseFailure
+          (Binary.label $ Proxy @(Annotator (TxWits era)))
+          (DeserialiseFailure 7 "Empty list of scripts is not allowed")
+      -- TODO replace this with `plutusV4Script` once that is added
+      it "8th field" . expectFailureOnEmptyField 8 $
+        DecoderErrorDeserialiseFailure
+          (Binary.label $ Proxy @(Annotator (TxWits era)))
+          (DeserialiseFailure 2 "An error occured while decoding (Int,Void) not a valid key:.\nError: 8")
 
 witsDuplicateVKeyWits :: Enc
 witsDuplicateVKeyWits =
@@ -138,38 +189,22 @@ witsDuplicatePlutusData =
   where
     dat = E @Int 0
 
-goldenListRedeemersDisallowed :: forall era. DijkstraEraTest era => Spec
-goldenListRedeemersDisallowed =
+goldenListRedeemersDisallowed :: forall era. DijkstraEraTest era => Version -> Spec
+goldenListRedeemersDisallowed version =
   it "Decoding Redeemers encoded as a list fails" $
     expectDecoderFailureAnn @(Redeemers era)
-      (eraProtVerLow @era)
+      version
       listRedeemersEnc
       ( DecoderErrorDeserialiseFailure
           "Annotator (MemoBytes (RedeemersRaw DijkstraEra))"
           (DeserialiseFailure 0 "List encoding of redeemers not supported starting with PV 12")
       )
 
-goldenDuplicateCertsDisallowed :: forall era. DijkstraEraTest era => Spec
-goldenDuplicateCertsDisallowed =
-  it "Decoding a transaction body with duplicate certificates fails" $
-    expectDecoderFailureAnn @(TxBody TopTx era)
-      version
-      (duplicateCertsTx @era version)
-      ( DecoderErrorDeserialiseFailure
-          "Annotator (MemoBytes (DijkstraTxBodyRaw TopTx DijkstraEra))"
-          ( DeserialiseFailure
-              143
-              "Final number of elements: 1 does not match the total count that was decoded: 2"
-          )
-      )
-  where
-    version = eraProtVerLow @era
-
-goldenDuplicateVKeyWitsDisallowed :: forall era. DijkstraEraTest era => Spec
-goldenDuplicateVKeyWitsDisallowed =
+goldenDuplicateVKeyWitsDisallowed :: forall era. DijkstraEraTest era => Version -> Spec
+goldenDuplicateVKeyWitsDisallowed version =
   it "Decoding a TxWits with duplicate VKeyWits fails" $
     expectDecoderFailureAnn @(TxWits era)
-      (eraProtVerLow @era)
+      version
       witsDuplicateVKeyWits
       ( DecoderErrorDeserialiseFailure
           "Annotator (MemoBytes (AlonzoTxWitsRaw DijkstraEra))"
@@ -179,8 +214,8 @@ goldenDuplicateVKeyWitsDisallowed =
           )
       )
 
-goldenDuplicateNativeScriptsDisallowed :: forall era. DijkstraEraTest era => Spec
-goldenDuplicateNativeScriptsDisallowed =
+goldenDuplicateNativeScriptsDisallowed :: forall era. DijkstraEraTest era => Version -> Spec
+goldenDuplicateNativeScriptsDisallowed version =
   it "Decoding a TxWits with duplicate native scripts fails" $
     expectDecoderFailureAnn @(TxWits era)
       version
@@ -189,19 +224,17 @@ goldenDuplicateNativeScriptsDisallowed =
           "Annotator"
           "Duplicate elements in the scripts Set were encountered"
       )
-  where
-    version = eraProtVerLow @era
 
 goldenDuplicatePlutusScriptsDisallowed ::
   forall era l.
   ( DijkstraEraTest era
   , EraPlutusTxInfo l era
   ) =>
-  SLanguage l -> Spec
-goldenDuplicatePlutusScriptsDisallowed slang =
+  Version -> SLanguage l -> Spec
+goldenDuplicatePlutusScriptsDisallowed version slang =
   it ("Decoding a TxWits with duplicate " <> show slang <> " scripts fails") $
     expectDecoderFailureAnn @(TxWits era)
-      (eraProtVerLow @era)
+      version
       (witsDuplicatePlutus @era slang)
       ( DecoderErrorDeserialiseFailure
           "Annotator (MemoBytes (AlonzoTxWitsRaw DijkstraEra))"
@@ -211,11 +244,11 @@ goldenDuplicatePlutusScriptsDisallowed slang =
           )
       )
 
-goldenDuplicatePlutusDataDisallowed :: forall era. DijkstraEraTest era => Spec
-goldenDuplicatePlutusDataDisallowed =
+goldenDuplicatePlutusDataDisallowed :: forall era. DijkstraEraTest era => Version -> Spec
+goldenDuplicatePlutusDataDisallowed version =
   it "Decoding a TxWits with duplicate plutus data fails" $
     expectDecoderFailureAnn @(TxWits era)
-      (eraProtVerLow @era)
+      version
       witsDuplicatePlutusData
       ( DecoderErrorCustom
           "Annotator"