+++ eras/mary/impl/test/Main.hs
@@ -3,6 +3,7 @@
 module Main where
 
 import Cardano.Ledger.Mary (MaryEra)
+import qualified Test.Cardano.Ledger.Allegra.Binary.Golden as Golden
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.JSON (roundTripJsonEraSpec)
 import qualified Test.Cardano.Ledger.Mary.Binary.CddlSpec as CddlSpec
@@ -23,3 +24,4 @@ main =
         Imp.spec @MaryEra
       roundTripJsonEraSpec @MaryEra
       roundTripJsonShelleyEraSpec @MaryEra
+      Golden.spec @MaryEra